Refactor model fields in mixins and entities
This commit refactors the model fields in the mixins and entities modules. It removes the 'created_by' and 'modified_by' fields from the ModelMixin class in the mixins module and replaces them with a 'creation_date' field. In the entities module, the __str__ method in the OrganizationApplication class is updated to return the 'creation_date' instead of the 'created_at' field.