feat: allow admin and creator to edit entity text (#1451)
Allow only creators or admins to edit entity texts

Enforce validation so that only the creator of an entity text or an
admin (staff) can update texts for organizations, groups, and events.

This ensures that entity texts cannot be modified by unauthorized users.

Add id as uuid4 format in those text models.

Add update method in views to make the validation.

Co-authored-by: Andrew Tavis McAllister <andrew.t.mcallister@gmail.com>